# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Predict Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterUserBody {
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  referredBy?: string | null;
}

export interface UpdateWalletBody {
  walletAddress: string;
}

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  points: number;
  totalEarned: number;
  isVip: boolean;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  referredBy?: string | null;
  loginStreak: number;
  /** @nullable */
  lastLoginDate?: string | null;
  createdAt: string;
}

export interface UserStats {
  totalPredictions: number;
  wins: number;
  losses: number;
  winRate: number;
  totalWagered: number;
  netProfit: number;
  referralCount: number;
  rank: number;
}

export type CreatePredictionBodyDirection =
  (typeof CreatePredictionBodyDirection)[keyof typeof CreatePredictionBodyDirection];

export const CreatePredictionBodyDirection = {
  long: "long",
  short: "short",
} as const;

export interface CreatePredictionBody {
  telegramId: string;
  direction: CreatePredictionBodyDirection;
  amount: number;
  entryPrice: number;
}

export interface ResolvePredictionBody {
  exitPrice: number;
}

export type PredictionStatus =
  (typeof PredictionStatus)[keyof typeof PredictionStatus];

export const PredictionStatus = {
  pending: "pending",
  won: "won",
  lost: "lost",
} as const;

export interface Prediction {
  id: number;
  telegramId: string;
  direction: string;
  amount: number;
  entryPrice: number;
  /** @nullable */
  exitPrice?: number | null;
  status: PredictionStatus;
  /** @nullable */
  payout?: number | null;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}

export interface LeaderboardEntry {
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  points: number;
  isVip: boolean;
  rank: number;
}

export interface Quest {
  id: number;
  title: string;
  description: string;
  reward: number;
  externalUrl: string;
  category: string;
  isVipOnly: boolean;
  iconName: string;
}

export interface ClaimQuestBody {
  telegramId: string;
}

export interface ClaimQuestResponse {
  pointsAwarded: number;
  newBalance: number;
  message: string;
}

export interface DailyRewardBody {
  telegramId: string;
}

export interface DailyRewardResponse {
  pointsAwarded: number;
  newBalance: number;
  streak: number;
  message: string;
  isVipBonus: boolean;
}

export type GetUserPredictionsParams = {
  limit?: number;
};

export type GetLeaderboardParams = {
  limit?: number;
};