# lib/api-zod/src/generated/types/claimQuestResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Predict Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimQuestResponse {
  pointsAwarded: number;
  newBalance: number;
  message: string;
}