# lib/api-zod/src/generated/types/createPredictionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Predict Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePredictionBodyDirection } from "./createPredictionBodyDirection";

export interface CreatePredictionBody {
  telegramId: string;
  direction: CreatePredictionBodyDirection;
  amount: number;
  entryPrice: number;
}