# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Predict Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  points: number;
  isVip: boolean;
  rank: number;
}