# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Predict Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionStatus } from "./predictionStatus";

export interface Prediction {
  id: number;
  telegramId: string;
  direction: string;
  amount: number;
  entryPrice: number;
  /** @nullable */
  exitPrice?: number | null;
  status: PredictionStatus;
  /** @nullable */
  payout?: number | null;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}