# lib/api-zod/src/generated/types/quest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Predict Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface Quest {
  id: number;
  title: string;
  description: string;
  reward: number;
  externalUrl: string;
  category: string;
  isVipOnly: boolean;
  iconName: string;
}