# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Predict Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  points: number;
  totalEarned: number;
  isVip: boolean;
  /** @nullable */
  walletAddress?: string | null;
  /** @nullable */
  referredBy?: string | null;
  loginStreak: number;
  /** @nullable */
  lastLoginDate?: string | null;
  createdAt: string;
}