# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alpha Predict Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  totalPredictions: number;
  wins: number;
  losses: number;
  winRate: number;
  totalWagered: number;
  netProfit: number;
  referralCount: number;
  rank: number;
}